feat: remove node, context & selection from event detail
BREAKING CHANGE: Ideally we'd extend the svg and canvas elements rather than wrap them but this has a number of issues (not least the custom element polyfills don't work for SVG). Previously we'd tried to hide that we'd wrapped the elements but on reflection this seems too magic so we've made this change to force users to understand the DOM structure we're actually creating.
The simple fix is to replace `selection` with `d3.select(nodes[i]/this).select('svg')` and `context` with or `nodes[i]/this.childNodes[0].getContext('2d')` (use `nodes[i]` if you're using fat arrows or `this` if you're using plain old `function`s)

BREAKING CHANGE: To remove the remaining magic we've also stopped
automatically propagating the data down to the wrapped element. If you use
the syntax suggested above this shouldn't have any effect.